Rename schemas to models
It can be confusing that the schema module created by diesel and the
schemas module have nearly the same name. Also schemas is not the best
naming now.
