+++ api/src/models/server.rs
